LearningAgent/gRPC-Server-Python/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
numpy==1.26.3
protobuf==4.25.2
pyrusgeom==0.1.2
scipy==1.11.4
